more updates to positioning to help prepare for eventual splitting
