+++ packages/office-ui-fabric-react/src/components/Callout/examples/Callout.Directional.Example.tsx
@@ -97,7 +97,6 @@ export class CalloutDirectionalExample extends React.Component<any, ICalloutDire
             beakWidth={ beakWidth }
             onDismiss={ this._onCalloutDismiss }
             directionalHint={ directionalHint }
-            onPositioned={ console.dir }
           >
             <div className='ms-CalloutExample-header'>
               <p className='ms-CalloutExample-title'>
